+++ rest-api/data/sites.csv
@@ -163,3 +163,8 @@ WALGREENS_PHOENIX,hpo-site-walgreensphoenixthomas,Thomas,DV Walgreens Phoenix,70
 WALGREENS_PHOENIX,hpo-site-walgreensphoenixapachejunction,Apache Junction,DV Walgreens Phoenix,7035735,,,Inactive,Inactive,Inactive,Inactive,Inactive,Inactive,Inactive,Inactive,Inactive,Inactive,Inactive,Inactive,,,,55 Apache Trail,,Apache Junction,AZ,85120,,,
 WALGREENS_PHOENIX,hpo-site-walgreensphoenix,Walgreens Phoenix,DV Walgreens Phoenix,7035735,,,Active,Inactive,Inactive,Active,Inactive,Inactive,Active,Inactive,Inactive,Active,Inactive,Inactive,,,,,,,,,,"jaime.metzler@walgreens.com, lucille.skuteris@walgreens.com",
 TEST_PROD,hpo-site-a,Prod HPO Site A,Prod HPO Site A,,,,Inactive,Inactive,Inactive,Inactive,Inactive,Inactive,Inactive,Inactive,Inactive,Inactive,Inactive,Inactive,,,,,,,,,,,
+TEST_PROD,hpo-site-a_test,Prod HPO Site A,Prod HPO Site A,,,,Inactive,Inactive,Inactive,Active,Active,Inactive,Inactive,Inactive,Inactive,Inactive,Inactive,Inactive,,,,,,,,,,,
+TEST_PROD,hpo-site-a_test_again,Prod HPO Site A,Prod HPO Site A,,,,Inactive,Inactive,Inactive,Active,Active,Inactive,Inactive,Inactive,Inactive,Inactive,Inactive,Inactive,,,,,,,,,,,
+TEST_PROD,hpo-site-a_test_again_agian,Prod HPO Site A,Prod HPO Site A,,,,Inactive,Inactive,Inactive,Active,Active,Inactive,Inactive,Inactive,Inactive,Inactive,Inactive,Inactive,,,,,,,,,,,
+TEST_PROD,hpo-site-a_test_2,Prod HPO Site A,Prod HPO Site A,,,,Inactive,Inactive,Inactive,Active,Active,Inactive,Inactive,Inactive,Inactive,Inactive,Inactive,Inactive,,,,,,,,,,,
+TEST_PROD,hpo-site-a_test_5,Prod HPO Site A,Prod HPO Site A,,,,Inactive,Inactive,Inactive,Active,Active,Inactive,Inactive,Inactive,Inactive,Inactive,Inactive,Inactive,,,,,,,,,,,
